Moved wrong iterator implementation rule to bugs
