+++ detekt-rules/src/main/kotlin/io/gitlab/arturbosch/detekt/rules/providers/ExceptionsProvider.kt
@@ -5,7 +5,6 @@ import io.gitlab.arturbosch.detekt.api.RuleSet
 import io.gitlab.arturbosch.detekt.api.RuleSetProvider
 import io.gitlab.arturbosch.detekt.rules.exceptions.ExceptionRaisedInUnexpectedLocation
 import io.gitlab.arturbosch.detekt.rules.exceptions.InstanceOfCheckForException
-import io.gitlab.arturbosch.detekt.rules.exceptions.IteratorNotThrowingNoSuchElementException
 import io.gitlab.arturbosch.detekt.rules.exceptions.PrintExceptionStackTrace
 import io.gitlab.arturbosch.detekt.rules.exceptions.RethrowCaughtException
 import io.gitlab.arturbosch.detekt.rules.exceptions.ReturnFromFinally
@@ -35,8 +34,7 @@ class ExceptionsProvider : RuleSetProvider {
 				ThrowingExceptionInMain(config),
 				RethrowCaughtException(config),
 				ThrowingNewInstanceOfSameException(config),
-				SwallowedException(config),
-				IteratorNotThrowingNoSuchElementException(config)
+				SwallowedException(config)
 		))
 	}
